pkg/validators/users.go
package validators

import (
	"errors"
	"unicode"
)

func ValidateLogin(login string) error {
	if len(login) == 0 {
		return errors.New("login cannot be empty")
	}

	if len(login) < 3 {
		return errors.New("login must be at least 3 characters long")
	}

	if len(login) > 20 {
		return errors.New("login must be no longer than 20 characters")
	}

	for _, char := range login {
		if !unicode.IsLetter(char) && !unicode.IsDigit(char) && char != '_' {
			return errors.New("login can only contain letters, digits, and underscores")
		}
	}

	return nil
}

func ValidatePassword(password string) error {
	if len(password) < 8 {
		return errors.New("password must be at least 8 characters long")
	}

	var hasUpper, hasLower, hasDigit, hasSpecial bool
	for _, char := range password {
		switch {
		case unicode.IsUpper(char):
			hasUpper = true
		case unicode.IsLower(char):
			hasLower = true
		case unicode.IsDigit(char):
			hasDigit = true
		case unicode.IsPunct(char) || unicode.IsSymbol(char):
			hasSpecial = true
		}
	}

	if !hasUpper {
		return errors.New("password must contain at least one uppercase letter")
	}
	if !hasLower {
		return errors.New("password must contain at least one lowercase letter")
	}
	if !hasDigit {
		return errors.New("password must contain at least one digit")
	}
	if !hasSpecial {
		return errors.New("password must contain at least one special character")
	}

	return nil
}
